refactor(decks): extract DeckButtonEditorViewModel from GridDeckViewModel
